fix(phpstan): drain the long tail — Str::replace, Category with(), custom-builder closures (#60)
Grab-bag of fixes for the one-offs still in the baseline:

- Str::replace overloads: when the search arg is an array the return
  type widens to iterable<string>|string. Wrap the three call sites in
  GenerateSocialCaption, CaptionGeneratorService with a (string) cast.

- Category::withActiveProducts / withFeaturedProducts scopes and the
  three callers (api/MenuController, storefront/MenuController,
  PrintableMenu) used arrow closures inside with() which PHPStan
  couldn't reconcile against the HasMany return. Swapped to regular
  non-returning closures — with() doesn't use the return value.

- CouponsTable filter closure: the inner scope was typed Builder<Coupon>
  but called valid() which lives on the custom CouponQueryBuilder.
  Retyped the @var cast to CouponQueryBuilder<Coupon>.

- CustomerDirectoryStatsQuery::get(): same shape — withSum closure was
  typed Builder, so $q->active() (Order scope) didn't resolve. Retyped
  as OrderQueryBuilder.

- PlatformStats / RevenueOverview: drop $t->plan->priceInDollars() ?? 0
  now that SubscriptionTier::priceInDollars() has a non-null int return
  (nullCoalesce.expr was flagging the redundant fallback).

- CheckoutController: abort_unless(SubscriptionTier::tryFrom(...)) fed
  the enum directly into a bool-typed arg; use tier !== null.

- AppIconGeneratorService: imagecreatetruecolor() expects int<1, max>,
  clamp size at the top with max(1, $size).

- TaxExport::generateExport(): was protected but called from the form
  schema closure; make it public to match its runtime use.

- ExportController::__invoke: annotate the spread array as
  array<int, string> so in_array narrowing satisfies abort_unless.

- CustomerDirectory::addNote passes Auth::id() (int|string|null) into an
  int-typed Action arg; cast to int.

- PriceSuggestionTool::generateMarginComparisons: selectedRecipe->cost
  is nullable; cast to float so PricingRecommendationService::suggestPrice
  still typechecks.

- CustomerPhotosTable bulk approve closure: import CustomerPhoto so
  the fn (CustomerPhoto $record) typehint resolves.

- ReviewQueryBuilder::statistics(): reconstruct the object explicitly
  from the first() result so the declared return shape is guaranteed.

- ProductAnalysisService: tighter return-type docs on formatIngredients
  (cast cost/quantity to float) and loadProductAnalysis (actually
  declare the row shape instead of array<string, mixed>).

- OrderIngredientAggregator::crossReferenceInventory: cast the max()
  deficit to float to match the declared shape.

- ActiveBakeriesQuery: ->values() after get()->map() to normalise the
  keys to Collection<int, …>, and cast the nullable $store_name to
  string to satisfy the array{name: string} shape.

- QrCodeService: the vendor returns its own HtmlString class that
  PHPStan can't resolve through a (string) cast. Extracted a private
  stringify() helper that uses instanceof / method_exists to normalise.

Baseline: 41 → 24. The remaining entries are mostly Larastan / vendor
edge cases.

+++ app/Services/Orders/OrderIngredientAggregator.php
@@ -97,7 +97,7 @@ class OrderIngredientAggregator
             $item['stock_unit'] = $tracked?->unit;
             $item['needs_purchase'] = $tracked ? $item['quantity'] > $currentStock : true;
             $item['deficit'] = $tracked
-                ? max(0, $item['quantity'] - $currentStock)
+                ? (float) max(0, $item['quantity'] - $currentStock)
                 : $item['quantity'];
 
             return $item;
